+++ changelog.md
@@ -1,1 +1,1 @@
-- Updated to 1.21.3
+- Added suggestion provider for function arguments